refactor(memory): neutral journal read-model + guard the churn scan (#434 review)
Address sage's two code findings on #428 (fixes completed by the agent before an
API drop; committed here):
- Layering: `computeNoteRetrievalCounts` (+ the shared #425 journal-streaming
  fold) moved out of `memory-audit.ts` into a neutral `src/memory-journal.ts`
  read-model consumed by BOTH audit and consolidation, so M6 no longer depends
  on the M7 audit module. Audit probe behaviour unchanged.
- Perf: `planAutoMergePairs` computes the assistant-eligible subset first and
  skips the O(events) journal scan when it's empty — a no-op consolidation no
  longer reads the journal.
